+++ coin-alert/package.json
@@ -20,7 +20,7 @@
     "chalk": "^5.4.1",
     "dotenv": "^16.4.7",
     "firebase": "^11.2.0",
-    "firebase-admin": "^13.0.2",
+    "firebase-admin": "^13.1.0",
     "next": "15.1.6",
     "next-auth": "^4.24.11",
     "react": "^19.0.0",
