+++ coin-alert/src/app/lib/utils/priceAlertHelper.ts
@@ -1,7 +1,7 @@
 import { TOKEN_PROGRAM_ID } from "@solana/spl-token";
 import { PublicKey } from "@solana/web3.js";
 import { connection } from "../connection";
-import { ALARM_CONFIGS_MAX, AlarmConfig, AlarmType, NOISIER_ALARM_CONFIGS, STANDARD_ALARM_CONFIGS } from "../constants/alarmConstants";
+import { AlarmConfig, AlarmType, NOISIER_ALARM_CONFIGS, STANDARD_ALARM_CONFIGS } from "../constants/alarmConstants";
 import { PriceData, Token } from "../firebase/tokenUtils";
 import { AlarmPreset } from "../firebase/userUtils";
 import { blockchainTaskQueue } from "../taskQueue";
@@ -44,13 +44,12 @@ export async function getLastHourPrices(token: Token | undefined): Promise<Price
 }
 
 
-export function getAlarmConfig(alarmPreset: AlarmPreset, userId: string){
+export function getAlarmConfig(alarmPreset: AlarmPreset){
     // console.log("Using alarm config: ")
     // console.log(NOISIER_ALARM_CONFIGS)
-    if(userId == "D7gDyfspTANknhFTJwYlCEM9NLW2"){
-        return ALARM_CONFIGS_MAX
-    }
-    throw Error("where is jeebs? userId = " + userId)
+    // if(userId == "D7gDyfspTANknhFTJwYlCEM9NLW2"){
+    //     return ALARM_CONFIGS_MAX
+    // }
     if(alarmPreset == "left"){
         return NOISIER_ALARM_CONFIGS
     }
